adds some more tests for the location spatial melting functions
